Rectification avec ajout des interfaces statiques du test technique

+++ recruit-me-ui/src/web/src/app/modules/recruteur/list-recruteurs/list-recruteurs.component.ts
@@ -3,6 +3,9 @@ import {Recruteur} from "../../../core/models/recruteur";
 import {RecruteurService} from "../../../core/services/recruteur-service/recruteur.service";
 import {Router} from "@angular/router";
 import {environment} from "../../../../environments/environment";
+import {MatDialog} from "@angular/material/dialog";
+import {MatSnackBar} from "@angular/material/snack-bar";
+import {YesOrNOComponent} from "../../../shared/yes-or-no/yes-or-no.component";
 
 @Component({
   selector: 'app-list-recruteurs',
@@ -14,7 +17,7 @@ export class ListRecruteursComponent implements OnInit {
   recruteur: Recruteur[];
 
   constructor(private recruteurService: RecruteurService,
-              private router: Router) { }
+              private router: Router, public dialog: MatDialog, private _snackBar: MatSnackBar) { }
 
   ngOnInit() {
     this.reloadData();
@@ -25,13 +28,26 @@ export class ListRecruteursComponent implements OnInit {
   }
 
   deleteRecruteur(id: number) {
-    this.recruteurService.deleteRecruteur(id)
-        .subscribe(
+    const nomUtilisateur = this.recruteur.filter(recruteur => recruteur.id === id)[0].nomUtilisateur;
+    const dialogRef = this.dialog.open(YesOrNOComponent, {
+      width: '250px',
+      data: {message: `Voulez vous supprimer ${nomUtilisateur} ?`}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.recruteurService.deleteRecruteur(id)
+          .subscribe(
             data => {
               console.log(data);
+              this._snackBar.open("utilisateur supprimé", "OK", {
+                duration: 2000,
+              });
               this.reloadData();
             },
             error => console.log(error));
+      }
+    });
   }
 
   //candidatDetails(id: number){
